seperated the modal into a different component
