+++ src/pages/PasscodeManagement.jsx
@@ -1,6 +1,5 @@
 "use client"
-
-import { useState, useEffect, useRef } from "react"
+import { useState, useEffect } from "react"
 import { useParams, useNavigate } from "react-router-dom"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import {
@@ -8,19 +7,18 @@ import {
   faKey,
   faTimes,
   faPlus,
-  faSave,
   faTrash,
   faEye,
   faEyeSlash,
   faSearch,
 } from "@fortawesome/free-solid-svg-icons"
 import { toast } from "react-toastify"
+import PasscodeModal from "../components/PasscodeModal"
 import "../stylesheets/passcodeManagement.css"
 
 const PasscodeManagement = () => {
   const { eventId } = useParams()
   const navigate = useNavigate()
-  const modalRef = useRef(null)
 
   const [passcodes, setPasscodes] = useState([])
   const [loading, setLoading] = useState(true)
@@ -55,23 +53,6 @@ const PasscodeManagement = () => {
 
     fetchPasscodes()
   }, [eventId])
-
-  // Close modal when clicking outside
-  useEffect(() => {
-    const handleClickOutside = (e) => {
-      if (modalRef.current && !modalRef.current.contains(e.target)) {
-        closeModal()
-      }
-    }
-
-    if (showModal) {
-      document.addEventListener("mousedown", handleClickOutside)
-    }
-
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside)
-    }
-  }, [showModal])
 
   const handleBack = () => {
     navigate(`/events/${eventId}`)
@@ -369,60 +350,16 @@ const PasscodeManagement = () => {
         </div>
       </div>
 
-      {/* Passcode Modal */}
-      {showModal && (
-        <div className="modal-overlay">
-          <div className="passcode-modal" ref={modalRef}>
-            <div className="modal-header">
-              <h2>
-                <FontAwesomeIcon icon={faKey} /> {editingId ? "Edit Passcode" : "Add New Passcode"}
-              </h2>
-              <button className="close-modal-button" onClick={closeModal}>
-                <FontAwesomeIcon icon={faTimes} />
-              </button>
-            </div>
-
-            <div className="modal-content">
-              <form className="passcode-form" onSubmit={handleSubmitForm}>
-                <div className={`form-group ${formErrors.name ? "has-error" : ""}`}>
-                  <label htmlFor="name">Name</label>
-                  <input
-                    type="text"
-                    id="name"
-                    name="name"
-                    value={formData.name}
-                    onChange={handleFormChange}
-                    placeholder="e.g., Front Desk Staff 1"
-                  />
-                  {formErrors.name && <div className="error-message">{formErrors.name}</div>}
-                </div>
-
-                <div className={`form-group ${formErrors.passcode ? "has-error" : ""}`}>
-                  <label htmlFor="passcode">Passcode</label>
-                  <input
-                    type="text"
-                    id="passcode"
-                    name="passcode"
-                    value={formData.passcode}
-                    onChange={handleFormChange}
-                    placeholder="Enter a numeric passcode"
-                  />
-                  {formErrors.passcode && <div className="error-message">{formErrors.passcode}</div>}
-                </div>
-
-                <div className="form-actions">
-                  <button type="button" className="cancel-button" onClick={closeModal}>
-                    <FontAwesomeIcon icon={faTimes} /> Cancel
-                  </button>
-                  <button type="submit" className="save-button">
-                    <FontAwesomeIcon icon={faSave} /> {editingId ? "Update" : "Save"} Passcode
-                  </button>
-                </div>
-              </form>
-            </div>
-          </div>
-        </div>
-      )}
+      {/* Use the PasscodeModal component instead of inline JSX */}
+      <PasscodeModal
+        showModal={showModal}
+        closeModal={closeModal}
+        formData={formData}
+        formErrors={formErrors}
+        handleFormChange={handleFormChange}
+        handleSubmitForm={handleSubmitForm}
+        editingId={editingId}
+      />
     </div>
   )
 }
