--- src/stylesheets/passcodeModal.css
@@ -1,258 +1,0 @@
-/* Passcode Modal Styles */
-.modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-  padding: 20px;
-}
-
-.passcode-modal {
-  background-color: white;
-  border-radius: 12px;
-  box-shadow: 0 10px 40px rgba(0, 0, 0, 0.2);
-  width: 100%;
-  max-width: 700px;
-  max-height: 90vh;
-  overflow-y: auto;
-  animation: modalFadeIn 0.3s;
-}
-
-@keyframes modalFadeIn {
-  from { opacity: 0; transform: translateY(20px); }
-  to { opacity: 1; transform: translateY(0); }
-}
-
-.modal-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 20px 24px;
-  border-bottom: 1px solid #eeeeee;
-}
-
-.modal-header h2 {
-  font-size: 20px;
-  color: #442777;
-  font-weight: 600;
-  margin: 0;
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.close-modal-button {
-  background: none;
-  border: none;
-  font-size: 18px;
-  color: #666;
-  cursor: pointer;
-  width: 36px;
-  height: 36px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 50%;
-  transition: background-color 0.2s;
-}
-
-.close-modal-button:hover {
-  background-color: #f6f4ff;
-  color: #442777;
-}
-
-.modal-content {
-  padding: 24px;
-}
-
-/* Passcode List */
-.passcode-list {
-  margin-bottom: 24px;
-}
-
-.passcode-list-header {
-  display: grid;
-  grid-template-columns: 1fr 1fr 100px;
-  gap: 16px;
-  padding: 10px 16px;
-  background-color: #f9f7ff;
-  border-radius: 8px;
-  font-size: 14px;
-  font-weight: 600;
-  color: #442777;
-  margin-bottom: 12px;
-}
-
-.passcode-item {
-  display: grid;
-  grid-template-columns: 1fr 1fr 100px;
-  gap: 16px;
-  padding: 12px 16px;
-  border-bottom: 1px solid #f0f0f0;
-  align-items: center;
-}
-
-.passcode-item:last-child {
-  border-bottom: none;
-}
-
-.passcode-name {
-  font-size: 15px;
-  font-weight: 500;
-  color: #333;
-}
-
-.passcode-value {
-  font-size: 15px;
-  color: #666;
-  font-family: monospace;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.toggle-visibility-button {
-  background: none;
-  border: none;
-  color: #6a5f8a;
-  cursor: pointer;
-  padding: 4px;
-  border-radius: 4px;
-}
-
-.toggle-visibility-button:hover {
-  background-color: #f6f4ff;
-  color: #442777;
-}
-
-.passcode-actions {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.edit-button {
-  background-color: transparent;
-  color: #442777;
-  border: 1px solid #442777;
-  border-radius: 6px;
-  padding: 4px 10px;
-  font-size: 12px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-.edit-button:hover {
-  background-color: #f6f4ff;
-}
-
-.delete-button {
-  background-color: transparent;
-  color: #ff4d4f;
-  border: 1px solid #ff4d4f;
-  border-radius: 6px;
-  padding: 4px 8px;
-  font-size: 12px;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-.delete-button:hover {
-  background-color: #fff1f0;
-}
-
-.no-passcodes-message {
-  text-align: center;
-  padding: 30px;
-  color: #666;
-  background-color: #f9f9f9;
-  border-radius: 8px;
-}
-
-/* Passcode Form */
-.add-passcode-button {
-  background-color: #442777;
-  color: white;
-  border: none;
-  border-radius: 8px;
-  padding: 10px 16px;
-  font-size: 15px;
-  font-weight: 500;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  margin: 0 auto;
-  transition: background-color 0.2s;
-}
-
-.add-passcode-button:hover {
-  background-color: #3a2065;
-}
-
-.passcode-form {
-  background-color: #f9f7ff;
-  border-radius: 8px;
-  padding: 20px;
-  margin-top: 20px;
-}
-
-.passcode-form h3 {
-  font-size: 16px;
-  font-weight: 600;
-  color: #442777;
-  margin: 0 0 16px 0;
-}
-
-.form-group {
-  margin-bottom: 16px;
-}
-
-.form-group label {
-  display: block;
-  margin-bottom: 6px;
-  font-size: 14px;
-  font-weight: 500;
-  color: #555;
-}
-
-.form-group input {
-  width: 100%;
-  padding: 10px 12px;
-  border-radius: 8px;
-  border: 1px solid #e0e0e0;
-  font-size: 14px;
-}
-
-.form-group input:focus {
-  outline: none;
-  border-color: #442777;
-  box-shadow: 0 0 0 2px rgba(68, 39, 119, 0.1);
-}
-
-/* Fix for media query that was cut off */
-@media (max-width: 480px) {
-  .passcode-list-header,
-  .passcode-item {
-    grid-template-columns: 1fr;
-  }
-  
-  .passcode-item {
-    padding-bottom: 50px;
-  }
-  
-  .passcode-value {
-    margin-bottom: 8px;
-  }
-  
-  .modal-content {
-    padding: 16px;
-  }
-}
